+++ cmaes_framework/snn_sim/run_simulation.py
@@ -67,10 +67,8 @@ def group_list(flat_list: list, n: int) -> list:
     """
     return [list(flat_list[i:i+n]) for i in range(0, len(flat_list), n)]
 
-def input_scaling_2(init_corners, cur_corners):
-    return ((init_corners - cur_corners) / init_corners)
 
-def run(iters, genome, mode, vid_name=None, vid_path=None):
+def run(iters, genome, mode, vid_name=None, vid_path=None, snn_logs=False):
     """
     Runs a single simulation of a given genome.
 
@@ -84,6 +82,7 @@ def run(iters, genome, mode, vid_name=None, vid_path=None):
                        "b: shows the simulation on a window and saves a video.
         vid_name (string): If mode is "v" or "b", this is the name of the saved video.
         vid_path (string): If mode is "v" or "b", this is the path the video will be saved.
+        snn_logs (bool): Whether to produce SNN logs.
     Returns:
         float: The fitness of the genome.
     """
@@ -122,9 +121,6 @@ def run(iters, genome, mode, vid_name=None, vid_path=None):
     snn_controller = SNNController(2, 2, 1, robot_config=robot_file_path)
     snn_controller.set_snn_weights(genome)
 
-    spike_trains = []
-    levels_log = []
-
     for i in range(iters):
 
         # Get point mass locations
@@ -133,15 +129,9 @@ def run(iters, genome, mode, vid_name=None, vid_path=None):
         # Get current corner distances
         corner_distances = np.array(morphology.get_corner_distances(raw_pm_pos))
 
-        if i == 0:
-            init_corner_distances = corner_distances
-
         # Use the normalized distances as input
-        action, spikes, levels = snn_controller.get_lengths(input_scaling_2(init_corner_distances, corner_distances))
-
-        spike_trains.append(spikes)
-        levels_log.append(levels)
-
+        action = snn_controller.get_lengths(corner_distances)
+  
         # Clip actuator target lengths to be between 0.6 and 1.6 to prevent buggy behavior
         action = np.clip(action, ACTUATOR_MIN_LEN, ACTUATOR_MAX_LEN)
 
@@ -170,4 +160,8 @@ def run(iters, genome, mode, vid_name=None, vid_path=None):
     if mode in ["v", "b"]:
         create_video(video_frames, vid_name, vid_path, FPS)
 
-    return (FITNESS_OFFSET - fitness), spike_trains, levels_log # Turn into a minimization problem
+    if snn_logs:
+        snn_controller.generate_output_csv()
+
+    return FITNESS_OFFSET - fitness # Turn into a minimization problem
+